refact: DOM 생성 함수를 widgets로 분리
